requirements.txt
autopep8==1.5.4
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.1.0
pycodestyle==2.6.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
